Fix IE7 not supporting data uri:S
We use the onerror event that will be trigered when trying to load a
data uri:s image on IE7, this is hacky but it works.

Untested under IE6 but he's DEAD right ?

We could also have used conditionnal comments and b.gif as a fallback
but that would really hurt the generated HTML which is already bigger
than original when using the lazyloader.